Move facet API endpoints to a controller.
