+++ includes/classes/Feature/Facets/Types/MetaRange/Block.php
@@ -23,7 +23,6 @@ class Block extends \ElasticPress\Feature\Facets\Block {
 	 */
 	public function setup() {
 		add_action( 'init', [ $this, 'register_block' ] );
-		add_action( 'rest_api_init', [ $this, 'setup_endpoints' ] );
 		add_action( 'enqueue_block_editor_assets', [ $this, 'enqueue_editor_assets' ] );
 		add_action( 'wp_enqueue_scripts', [ $this, 'enqueue_assets' ] );
 	}
@@ -76,51 +75,6 @@ class Block extends \ElasticPress\Feature\Facets\Block {
 	}
 
 	/**
-	 * Setup REST endpoints for the feature.
-	 *
-	 * @return void
-	 */
-	public function setup_endpoints() {
-		register_rest_route(
-			'elasticpress/v1',
-			'facets/meta-range/keys',
-			[
-				'methods'             => 'GET',
-				'permission_callback' => [ $this, 'check_facets_rest_permission' ],
-				'callback'            => [ $this, 'get_rest_registered_metakeys' ],
-			]
-		);
-
-		register_rest_route(
-			'elasticpress/v1',
-			'facets/meta-range/block-preview',
-			[
-				'methods'             => 'GET',
-				'permission_callback' => [ $this, 'check_facets_rest_permission' ],
-				'callback'            => [ $this, 'render_block_preview' ],
-				'args'                => [
-					'facet' => [
-						'sanitize_callback' => 'sanitize_text_field',
-					],
-				],
-
-			]
-		);
-	}
-
-	/**
-	 * DEPRECATED. Check permissions of the /facets/meta-range/* REST endpoints.
-	 *
-	 * @deprecated 4.7.0
-	 * @return WP_Error|true
-	 */
-	public function check_facets_meta_rest_permission() {
-		_deprecated_function( __FUNCTION__, '4.7.0', '$this->check_facets_rest_permission()' );
-
-		return $this->check_facets_rest_permission();
-	}
-
-	/**
 	 * Render the block.
 	 *
 	 * @param array $attributes Block attributes.
@@ -140,88 +94,5 @@ class Block extends \ElasticPress\Feature\Facets\Block {
 		</div>
 		<?php
 		return ob_get_clean();
-	}
-
-
-	/**
-	 * Outputs the block preview
-	 *
-	 * @param \WP_REST_Request $request REST request
-	 * @return string
-	 */
-	public function render_block_preview( $request ) {
-		global $wp_query;
-
-		add_filter( 'ep_is_facetable', '__return_true' );
-
-		$search = \ElasticPress\Features::factory()->get_registered_feature( 'search' );
-		$facets = \ElasticPress\Features::factory()->get_registered_feature( 'facets' );
-
-		$facet = $request->get_param( 'facet' );
-
-		add_filter(
-			'ep_facet_meta_range_fields',
-			function ( $meta_fields ) use ( $facet ) {
-				$meta_fields = [ $facet ];
-
-				return $meta_fields;
-			}
-		);
-
-		$args = [
-			'post_type'      => $search->get_searchable_post_types(),
-			'posts_per_page' => 1,
-		];
-		$wp_query->query( $args );
-
-		$min_field_name = $facets->types['meta-range']->get_filter_name() . $facet . '_min';
-		$max_field_name = $facets->types['meta-range']->get_filter_name() . $facet . '_max';
-
-		if ( empty( $GLOBALS['ep_facet_aggs'][ $min_field_name ] ) || empty( $GLOBALS['ep_facet_aggs'][ $max_field_name ] ) ) {
-			return wp_send_json_error();
-		}
-
-		$min = $GLOBALS['ep_facet_aggs'][ $min_field_name ];
-		$max = $GLOBALS['ep_facet_aggs'][ $max_field_name ];
-
-		return wp_send_json_success(
-			[
-				'min' => $min,
-				'max' => $max,
-			]
-		);
-	}
-
-	/**
-	 * Return an array of registered meta keys.
-	 *
-	 * @return array
-	 */
-	public function get_rest_registered_metakeys() {
-		$post_indexable = \ElasticPress\Indexables::factory()->get( 'post' );
-
-		try {
-			$meta_keys = $post_indexable->get_distinct_meta_field_keys();
-		} catch ( \Throwable $th ) {
-			$meta_keys = [];
-		}
-
-		return $meta_keys;
-	}
-
-	/**
-	 * Utilitary method to set default attributes.
-	 *
-	 * @param array $attributes Attributes passed
-	 * @return array
-	 */
-	protected function parse_attributes( $attributes ) {
-		_doing_it_wrong(
-			__METHOD__,
-			esc_html__( 'Attribute parsing is now left to block.json.', 'elasticpress' ),
-			'4.7.0'
-		);
-
-		return $attributes;
 	}
 }
